Feature/PAT-673/project contributor can only modify their tenders

Approved-by: Daniel Ashworth

+++ src/main/java/uk/co/ogauthority/pathfinder/controller/project/upcomingtender/UpcomingTendersController.java
@@ -28,7 +28,9 @@ import uk.co.ogauthority.pathfinder.controller.project.annotation.ProjectFormPag
 import uk.co.ogauthority.pathfinder.controller.project.annotation.ProjectStatusCheck;
 import uk.co.ogauthority.pathfinder.controller.project.annotation.ProjectTypeCheck;
 import uk.co.ogauthority.pathfinder.controller.rest.TenderFunctionRestController;
+import uk.co.ogauthority.pathfinder.exception.AccessDeniedException;
 import uk.co.ogauthority.pathfinder.model.entity.project.ProjectDetail;
+import uk.co.ogauthority.pathfinder.model.entity.project.upcomingtender.UpcomingTender;
 import uk.co.ogauthority.pathfinder.model.enums.ValidationType;
 import uk.co.ogauthority.pathfinder.model.enums.audit.AuditEvent;
 import uk.co.ogauthority.pathfinder.model.enums.project.ContractBand;
@@ -157,6 +159,7 @@ public class UpcomingTendersController extends PathfinderFileUploadController {
                                          @PathVariable("upcomingTenderId") Integer upcomingTenderId,
                                          ProjectContext projectContext) {
     var upcomingTender = upcomingTenderService.getOrError(upcomingTenderId);
+    checkIfUserHasAccessToTender(upcomingTender);
     return getUpcomingTenderModelAndView(
         projectContext.getProjectDetails(),
         upcomingTenderService.getForm(upcomingTender)
@@ -171,6 +174,7 @@ public class UpcomingTendersController extends PathfinderFileUploadController {
                                            ValidationType validationType,
                                            ProjectContext projectContext) {
     var upcomingTender = upcomingTenderService.getOrError(upcomingTenderId);
+    checkIfUserHasAccessToTender(upcomingTender);
     bindingResult = upcomingTenderService.validate(form, bindingResult, validationType);
     return controllerHelperService.checkErrorsAndRedirect(
         bindingResult,
@@ -197,7 +201,7 @@ public class UpcomingTendersController extends PathfinderFileUploadController {
                                                   @PathVariable("displayOrder") Integer displayOrder,
                                                   ProjectContext projectContext) {
     var upcomingTender = upcomingTenderService.getOrError(upcomingTenderId);
-
+    checkIfUserHasAccessToTender(upcomingTender);
     var modelAndView = new ModelAndView("project/upcomingtender/removeUpcomingTender")
           .addObject("view", upcomingTenderSummaryService.getUpcomingTenderView(upcomingTender, displayOrder))
           .addObject("cancelUrl", ReverseRouter.route(on(UpcomingTendersController.class).viewUpcomingTenders(projectId, null)));
@@ -211,6 +215,7 @@ public class UpcomingTendersController extends PathfinderFileUploadController {
                                            @PathVariable("displayOrder") Integer displayOrder,
                                            ProjectContext projectContext) {
     var upcomingTender = upcomingTenderService.getOrError(upcomingTenderId);
+    checkIfUserHasAccessToTender(upcomingTender);
     upcomingTenderService.delete(upcomingTender);
     AuditService.audit(
         AuditEvent.UPCOMING_TENDER_REMOVED,
@@ -301,5 +306,15 @@ public class UpcomingTendersController extends PathfinderFileUploadController {
         .addObject("contractBands", ContractBand.getAllAsMap());
     breadcrumbService.fromUpcomingTenders(projectDetail.getProject().getId(), modelAndView, PAGE_NAME_SINGULAR);
     return modelAndView;
+  }
+
+  private void checkIfUserHasAccessToTender(UpcomingTender upcomingTender) {
+    if (!upcomingTenderService.canCurrentUserAccessTender(upcomingTender)) {
+      throw new AccessDeniedException(
+          String.format(
+              "User does not have access to update UpcomingTender with id: %d",
+              upcomingTender.getId())
+      );
+    }
   }
 }
